Stop injecting whitespaces into extracted text
The function setSortByPosition() injects whitespaces into the text. This complicates the regex to parse the text. I am trying to add a function to import a "Buchungsanzeige" from comdirect and the spaces appear in the names of the options, so it's impossible to get the names correct. The number of failed tests did not increase for me after changing it to false, since in the regex the number of whitespaces between the letters usually is optional. This change should allow making the code of the extractors a lot simpler.